requirements.txt
qdrant-client[fastembed]
opencv-python
streamlit
watchdog
pandas
numpy
SpeechRecognition
pydub
Pillow
